consolidated string dict to map functions
