+++ src/arr/compiler/ts-codegen-helpers.ts
@@ -1,5 +1,5 @@
 import * as J from 'estree';
-import type { List } from './ts-impl-types';
+import type { List, MutableStringDict, StringDict, PMethod, PFunction } from './ts-impl-types';
 import type * as A from './ts-ast';
 
 export type Variant<T, V> = T & { $name: V };
@@ -16,6 +16,23 @@ type PyretDataValue = {
     [property: string]: any
   }
 };
+
+type DropFirst<T extends unknown[]> = ((...p: T) => void) extends ((p1: infer P1, ...rest: infer R) => void) ? R : never
+
+type SDExports = {
+  dict: { values: { dict: {
+    'make-mutable-string-dict': PFunction<<T>() => MutableStringDict<T>>
+    'is-mutable-string-dict': PFunction<<T>(val: any) => val is MutableStringDict<T>>,
+    'make-string-dict': PFunction<<T>() => StringDict<T>>,
+    'is-string-dict': PFunction<<T>(val: any) => val is StringDict<T>>,
+    'map-keys': PFunction<<T, U>(f: ((key: T) => U), isd: StringDict<T>) => List<U>>,
+    'map-keys-now': PFunction<<T, U>(f: ((key: T) => U), msd: MutableStringDict<T>) => List<U>>,
+    'fold-keys': PFunction<<T, U>(f: (key: string, acc: U) => U, init: U, isd: StringDict<T>) => U>,
+    'fold-keys-now': PFunction<<T, U>(f: (key: string, acc: U) => U, init: U, msd: MutableStringDict<T>) => U>,
+    'each-key': PFunction<<T>(f: ((key: T) => void), isd: StringDict<T>) => void>,
+    'each-key-now': PFunction<<T>(f: ((key: T) => void), msd: MutableStringDict<T>) => void>,
+  }}}
+}
 
 export type Visitor<T extends PyretDataValue, Ret = any> = {
   [method in T["$name"]]?: (self: Visitor<T, Ret>, val: Variant<T, method>) => Ret;
@@ -82,10 +99,16 @@ export interface Exports {
       v : Visitor<T>,
       d : T,
     ) => Ret,
+  callMethod: <Name extends string, O extends {dict: {[n in Name]: PMethod<any, (...args: any[]) => any>}}>(obj : O, name: Name, ...args: DropFirst<Parameters<O["dict"][Name]["full_meth"]>>) => ReturnType<O["dict"][Name]["full_meth"]>,
+  mapFromStringDict: <T>(s : StringDict<T>) => Map<string, T>,
+  mapFromMutableStringDict: <T>(s : MutableStringDict<T>) => Map<string, T>,
+  stringDictFromMap: <T>(m : Map<string, T>) => StringDict<T>,
+  mutableStringDictFromMap: <T>(m : Map<string, T>) => MutableStringDict<T>,
 }
 
 ({
   requires: [
+    { 'import-type': 'builtin', name: 'string-dict' },
     { 'import-type': 'dependency', protocol: 'file', args: ['ast.arr']},
   ],
   nativeRequires: [],
@@ -131,7 +154,7 @@ export interface Exports {
       compileSrcloc: 'tany',
     },
   },
-  theModule: function(runtime, _, __, Ain: A.Exports) {
+  theModule: function(runtime, _, __, SD: SDExports, Ain: A.Exports) {
     const A = Ain.dict.values.dict;
     class ExhaustiveSwitchError extends Error {
       constructor(v: never, message?: string) {
@@ -538,7 +561,33 @@ export interface Exports {
       }
     }
 
-
+    function callMethod<Name extends string, O extends {dict: {[n in Name]: PMethod<any, (...args: any[]) => any>}}>(obj : O, name: Name, ...args: DropFirst<Parameters<O["dict"][Name]["full_meth"]>>) : ReturnType<O["dict"][Name]["full_meth"]> {
+      return obj.dict[name].full_meth(obj, ...args);
+    }
+    function mapFromStringDict<T>(s : StringDict<T>) : Map<string, T> {
+      const m : Map<string, T> = new Map();
+      for (let valKey of listToArray(callMethod(s, 'keys-list'))) {
+        m.set(valKey, callMethod(s, "get-value", valKey));
+      }
+      return m;
+    }
+    function mapFromMutableStringDict<T>(s : MutableStringDict<T>) : Map<string, T> {
+      const m : Map<string, T> = new Map();
+      for (let valKey of listToArray(callMethod(s, 'keys-list-now'))) {
+        m.set(valKey, callMethod(s, "get-value-now", valKey));
+      }
+      return m;
+    }
+    function stringDictFromMap<T>(m : Map<string, T>): StringDict<T> {
+      return callMethod(mutableStringDictFromMap(m), 'freeze');
+    }
+    function mutableStringDictFromMap<T>(m : Map<string, T>): MutableStringDict<T> {
+      const s = SD.dict.values.dict['make-mutable-string-dict'].app<T>();
+      for (const [k, v] of m.entries()) {
+        callMethod(s, 'set-now', k, v);
+      }
+      return s;
+    }
 
     return runtime.makeJSModuleReturn({
       ArrayExpression,
@@ -583,6 +632,11 @@ export interface Exports {
       formatSrcloc,
       visit,
       map,
+      callMethod,
+      mapFromStringDict,
+      mapFromMutableStringDict,
+      stringDictFromMap,
+      mutableStringDictFromMap,
     });
   }
 })
